+++ tasks/git_changelog_generate.js
@@ -23,6 +23,7 @@ Changelog.prototype.parseRawCommit = require('./lib/parse-raw-commit');
 Changelog.prototype.linkToIssue = require('./lib/link-to-issue');
 Changelog.prototype.linkToCommit = require('./lib/link-to-commit');
 Changelog.prototype.currentDate = require('./lib/current-date');
+Changelog.prototype.printHeader = require('./lib/print-header');
 Changelog.prototype.printSection = require('./lib/print-section');
 Changelog.prototype.printSalute = require('./lib/print-salute');
 Changelog.prototype.readGitLog = require('./lib/read-gitlog');