+++ test/git_changelog_generate.spec.js
@@ -101,6 +101,12 @@ describe('git_changelog_generate.js', function() {
         expect(changelog.options.msg).to.contain('file: test');
       });
 
+      it('should store "logo" if passed as an option', function() {
+        changelog.initOptions({ logo: 'test' });
+        expect(changelog.options.logo).to.equal('test');
+        expect(changelog.options.msg).to.contain('logo: test');
+      });
+
       it('should store "grep_commits" if passed as an option', function() {
         changelog.initOptions({ grep_commits: 'test' });
         expect(changelog.options.grep_commits).to.equal('test');
@@ -530,6 +536,70 @@ describe('git_changelog_generate.js', function() {
           expect(changelog.printSalute).to.have.been.calledWith(this.stream);
         });
 
+      });
+
+      describe('with logo', function() {
+
+        before(function(done) {
+          this.stream = {
+            write: sinon.stub(),
+            end : function(){
+
+            },
+            on: sinon.spy(function(event, callback) {
+              callback();
+            })
+          };
+          this.commits = require('./fixtures/commits.js').withoutBreaking;
+
+          sinon.stub(changelog, 'organizeCommits');
+          sinon.stub(changelog, 'printSalute');
+          sinon.stub(changelog, 'printSection');
+
+          changelog.initOptions({ app_name: 'app', version: 'version' });
+          changelog.writeChangelog(this.stream, this.commits).then(function() {
+            done();
+          });
+        });
+
+        after(function() {
+          changelog.organizeCommits.restore();
+          changelog.printSection.restore();
+          changelog.printSalute.restore();
+        });
+
+        it('should organize commits', function() {
+          expect(changelog.organizeCommits).to.have.been.calledOnce;
+          expect(changelog.organizeCommits).to.have.been.calledWith(this.commits);
+        });
+
+        it('should write the header to the stream', function() {
+          expect(this.stream.write).to.have.been.calledOnce;
+          expect(this.stream.write).to.have.been.calledWithMatch(/^<a name=/);
+        });
+
+        it('should print 7 sections', function() {
+          var sections = [
+            'Bug Fixes',
+            'Features',
+            'Refactor',
+            'Style',
+            'Test',
+            'Chore',
+            'Documentation'
+          ];
+
+          expect(changelog.printSection.callCount).to.equal(7);
+          sections.forEach(function(section, index) {
+            var call = changelog.printSection.getCall(index);
+            expect(call.args).to.include(section);
+          });
+        });
+
+        it('should print salute', function() {
+          expect(changelog.printSalute).to.have.been.calledOnce;
+          expect(changelog.printSalute).to.have.been.calledWith(this.stream);
+        });
       });
 
     });