+++ CHANGELOG.md
@@ -1,5 +1,21 @@
-<a name="">my name</a>
-#  (2015-04-21)
+
+
+---
+<a name="">Version </a>
+<img width="300px" src="http://upload.wikimedia.org/wikipedia/commons/4/4a/Logo_2013_Google.png" />
+__null my name__
+_a project by Google Corp_
+#  v (2015-06-04)
+
+## Bug Fixes
+
+- **checkPath:** add missing require('path')
+  ([e5dab826](https://github.com/rafinskipg/git-changelog/commit/e5dab826062bd22dd37c8c3d3c24a4d9b4701f6d))
+
+
+## Documentation
+
+- **README:** add logo